requirements.txt
fastapi
uvicorn
paho-mqtt
psycopg2-binary
sqlalchemy
